wine-educator/requirements.txt
openai>=1.12.0
pinecone>=3.0.0
python-dotenv==1.0.0
streamlit>=1.31.0